+++ test/libsolidity/syntaxTests/nameAndTypeResolution/200_integer_signed_exp_unsigned.sol
@@ -1,3 +1,5 @@
-contract test { function() external { uint x = 3; int y = -4; y ** x; } }
-// ----
-// TypeError: (62-68): Operator ** not compatible with types int256 and uint256
+contract test {
+    function() external { uint x = 3; int y = -4; y ** x; }
+    function f() public pure { int16 x = 3; uint8 y = 4; x ** y; }
+    function g() public pure { int16 x = 3; uint16 y = 4; x ** y; }
+}