+++ Changelog.md
@@ -2,6 +2,7 @@
 
 Breaking changes:
  * General: Disallow explicit conversions from external function types to ``address`` and add a member called ``address`` to them as replacement.
+ * Type checker: Resulting type of exponentiation is equal to the type of the base. Also allow signed types for the base.
 
 
 Language Features: